+++ tests/PathOfExile.GameClient.Monitor.Tests/Given_DefaultNotificationMatching/GivenDefaultNotificationMatching.cs
@@ -36,6 +36,11 @@ namespace PathOfExile.GameClient.Monitor.Tests.Given_DefaultNotificationMatching
             NotificationMatch = DefaultNotificationMatching.PlayerInAreaNotificationMatch;
         }
 
+        protected void With_KitavasAfflictionNotificationMatch()
+        {
+            NotificationMatch = DefaultNotificationMatching.KitavasAfflictionNotificationMatch;
+        }
+
         protected void With_TradeNotificationMatch()
         {
             NotificationMatch = DefaultNotificationMatching.TradeNotificationMatch;