+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ##### features
 - Implemented matching for trade templates based on poe.trade and map exchange services
+- Implemented matching for Kitavas Affliction (resistatnce penalty)
 
 ## 0.1.0a